Changed orb package name and reorganized some files.
