+++ pkg/controllers/orb/proto/ORBLogEntry.proto
